+++ src/zustand/settings/LangSettingsZustand.ts
@@ -14,8 +14,9 @@ const zustandLsName = 'zustandLangSettings'
 
 
 
+const recoilLsName = 'langSettings'
 // To trigger Zustand update from Recoil to Zustand
-if (notExists(localStorage.getItem(zustandLsName))) {
+if (notExists(localStorage.getItem(zustandLsName)) && exists(recoilLsName)) {
   localStorage.setItem(zustandLsName, JSON.stringify({ version: -1 }))
 }
 
@@ -41,7 +42,6 @@ export const useLangSettingsZustand = create<LangSettingsZustand>()(persist(
     version: 1,
     migrate: (persisted: any, persistedVersion) => {
       if (persistedVersion <= 0) {
-        const recoilLsName = 'langSettings'
         const oldRaw = localStorage.getItem(recoilLsName)
         localStorage.removeItem(recoilLsName)
         const old = exists(oldRaw) ? JSON.parse(oldRaw) : undefined
